requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
PyYAML
python-box
psycopg2-binary
SQLAlchemy
python-dotenv